+++ src/main/java/com/evacipated/cardcrawl/modthespire/GithubUpdateChecker.java
@@ -1,5 +1,9 @@
 package com.evacipated.cardcrawl.modthespire;
 
+import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
@@ -26,5 +30,31 @@ public class GithubUpdateChecker extends UpdateChecker
     public URL getLatestReleaseURL() throws IOException
     {
         return new URL(getElementAsString("html_url"));
+    }
+
+    @Override
+    public URL getLatestDownloadURL() throws IOException
+    {
+        JsonElement assets = getElement("assets");
+
+        if (!(assets instanceof JsonArray)) {
+            throw new RuntimeException("Excepted assets to be of type array");
+        }
+
+        for (JsonElement asset : (JsonArray)assets) {
+            if (asset instanceof JsonObject) {
+                JsonObject object = (JsonObject)asset;
+                String url = object.get("browser_download_url").getAsString();
+                if (isJar(url)) {
+                    return new URL(url);
+                }
+            }
+        }
+
+        throw new RuntimeException("Could not find jar asset to download");
+    }
+
+    private boolean isJar(String url) {
+        return url.toLowerCase().endsWith(".jar");
     }
 }