+++ src/main/java/com/evacipated/cardcrawl/modthespire/LoadOrder.java
@@ -13,6 +13,8 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 
 import com.evacipated.cardcrawl.modthespire.lib.ConfigUtils;
+import com.evacipated.cardcrawl.modthespire.ui.JModPanelCheckBoxList;
+import com.evacipated.cardcrawl.modthespire.ui.ModPanel;
 import org.w3c.dom.Document;
 import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
@@ -25,27 +27,27 @@ public class LoadOrder {
     private static class ModDescriptor {
         public File mod;
         public ModInfo info;
-		public boolean checked;
+        public boolean checked;
         
         public ModDescriptor(File mod, ModInfo info, boolean checked) {
             this.mod = mod;
             this.info = info;
-			this.checked = checked;
+            this.checked = checked;
         }
     }
     
-    public static void defaultLoad(DefaultListModel<ModPanel> model, File[] mods, ModInfo[] info, Dimension parentSize) {
+    public static void defaultLoad(DefaultListModel<ModPanel> model, File[] mods, ModInfo[] info, Dimension parentSize, JModPanelCheckBoxList parent) {
         for (int i = 0; i < info.length; i++) {
-            model.addElement(new ModPanel(info[i], mods[i], parentSize));
+            model.addElement(new ModPanel(info[i], mods[i], parentSize, parent));
         }
         return;
     }
     
-    public static void loadModsInOrder(DefaultListModel<ModPanel> model, File[] mods, ModInfo[] info, Dimension parentSize) {
+    public static void loadModsInOrder(DefaultListModel<ModPanel> model, File[] mods, ModInfo[] info, Dimension parentSize, JModPanelCheckBoxList parent) {
         File cfg_file = new File(CFG_FILE);
         
         if (!cfg_file.exists()) {
-            defaultLoad(model, mods, info, parentSize);
+            defaultLoad(model, mods, info, parentSize, parent);
             return;
         }
         
@@ -57,7 +59,7 @@ public class LoadOrder {
             System.out.println("could not load config file: " + CFG_FILE);
             System.out.println("exception was: " + e.toString());
             e.printStackTrace();
-            defaultLoad(model, mods, info, parentSize);
+            defaultLoad(model, mods, info, parentSize, parent);
             return;
         }
         
@@ -97,8 +99,8 @@ public class LoadOrder {
         
         // actually set them in order in the list
         for (ModDescriptor descriptor : loadOrder) {
-			ModPanel toAdd = new ModPanel(descriptor.info, descriptor.mod, parentSize);
-			if (toAdd.checkBox.isEnabled()) {
+            ModPanel toAdd = new ModPanel(descriptor.info, descriptor.mod, parentSize, parent);
+            if (toAdd.checkBox.isEnabled()) {
                 toAdd.checkBox.setSelected(descriptor.checked);
             }
             model.addElement(toAdd);