+++ src/main/java/com/evacipated/cardcrawl/modthespire/ModUpdate.java
@@ -1,0 +1,17 @@
+package com.evacipated.cardcrawl.modthespire;
+
+import java.net.URL;
+
+public class ModUpdate
+{
+    public ModInfo info;
+    public URL releaseURL;
+    public URL downloadURL;
+
+    ModUpdate(ModInfo info, URL releaseURL, URL downloadURL)
+    {
+        this.info = info;
+        this.releaseURL = releaseURL;
+        this.downloadURL = downloadURL;
+    }
+}