+++ src/main/java/com/evacipated/cardcrawl/modthespire/patcher/LocatorInfo.java
@@ -4,18 +4,18 @@ import com.evacipated.cardcrawl.modthespire.lib.SpireInsertLocator;
 import javassist.CtBehavior;
 
 public class LocatorInfo {
-	
-	private CtBehavior ctMethodToPatch;
-	private Class<?> finderClass;
-	
-	public LocatorInfo(CtBehavior ctMethodToPatch, Class<?> finderClass) {
-		this.ctMethodToPatch = ctMethodToPatch;
-		this.finderClass = finderClass;
-	}
-	
-	public int[] findLines() throws Exception {
+
+    private CtBehavior ctMethodToPatch;
+    private Class<?> finderClass;
+
+    public LocatorInfo(CtBehavior ctMethodToPatch, Class<?> finderClass) {
+        this.ctMethodToPatch = ctMethodToPatch;
+        this.finderClass = finderClass;
+    }
+
+    public int[] findLines() throws Exception {
         SpireInsertLocator obj = (SpireInsertLocator)finderClass.newInstance();
         return obj.Locate(ctMethodToPatch);
-	}
+    }
 
 }